aba de procura e dashboard

+++ Y/app/routes.py
@@ -1,7 +1,7 @@
 print('começando routes')
 from flask import render_template, url_for, flash, redirect, request
 from app import app, db
-from app.forms import Register, LoginForm, PostForm
+from app.forms import Register, LoginForm, PostForm, SearchForm
 from app.models import Users, Posts
 from werkzeug.security import generate_password_hash, check_password_hash
 from flask_login import login_required, login_user, LoginManager, logout_user, current_user
@@ -36,15 +36,14 @@ def register():
             user = Users(username=form.username.data, name=form.name.data, email=form.email.data, password_hash=hashed_password)
             db.session.add(user)
             db.session.commit()
-            print('User submited!')
+            print('Registration Successful! Please Log In.')
+            return redirect(url_for('login'))
         name = form.name.data
         password_hash = form.password_hash.data
         form.username.data = ''
         form.name.data = ''
         form.email.data = ''
         form.password_hash.data = ''
-        flash("Registration Form Submitted Successfully!")
-        # colocar um redirect aqui para login
     our_users = Users.query.order_by(Users.date_added)
     return render_template('register.html', name = name, password_hash = password_hash, form = form, our_users = our_users)
 
@@ -66,7 +65,7 @@ def login():
             if check_password_hash(user.password_hash, form.password.data):
                 login_user(user)
                 print('login sucessful')
-                return redirect(url_for('home'))
+                return redirect(url_for('posts'))
             else:
                 print('wrong password')
         else:
@@ -82,46 +81,67 @@ def logout():
     print('logout sucedido')
     return redirect(url_for('login'))
 
-@app.route('/dashboard')
+@app.route('/dashboard', methods=['GET', 'POST'])
 @login_required
 def dashboard():
-    return render_template('dashboard.html')
+    form = Register()
+    id = current_user.id
+    name_to_update = Users.query.get_or_404(id)
+    if request.method == "POST":
+        name_to_update.name = request.form.get('name')
+        name_to_update.username = request.form.get('username')
+        name_to_update.email = request.form.get('email')
+        try:
+            db.session.commit()
+            flash("User Updated Successfully!")
+            return render_template('dashboard.html', name_to_update = name_to_update, form = form)
+        except:
+            flash("Whoops! There Was A Problem Updating The User!")
+            return render_template('dashboard.html', name_to_update = name_to_update, form = form)
+    else:
+        return render_template('dashboard.html', name_to_update = name_to_update, form = form)
 
-# Create User Profile Page
-@app.route('/user_profile/<int:id>', methods=['GET', 'POST'])
-def user_profile(id):
+# Create Update User Profile Page
+@app.route('/update_profile/<int:id>', methods=['GET', 'POST'])
+def update_profile(id):
     form = Register()
     name_to_update = Users.query.get_or_404(id)
     if request.method == "POST":
         name_to_update.name = request.form.get('name')
+        name_to_update.username = request.form.get('username')
         name_to_update.email = request.form.get('email')
-        name_to_update.password = request.form.get('password')  
+        name_to_update.password_hash = request.form.get('password_hash')  
         try:
             db.session.commit()
             flash("User Updated Successfully!")
-            return render_template('user_profile.html', name_to_update = name_to_update, form = form)
+            return render_template('update_profile.html', name_to_update = name_to_update, form = form)
         except:
             flash("Whoops! There Was A Problem Updating The User!")
-            return render_template('user_profile.html', name_to_update = name_to_update, form = form)
+            return render_template('update_profile.html', name_to_update = name_to_update, form = form)
     else:
-        return render_template('user_profile.html', name_to_update = name_to_update, form = form)
+        return render_template('update_profile.html', name_to_update = name_to_update, form = form)
 
 # Create Delete User Page
 @app.route('/delete/<int:id>')
+@login_required
 def delete(id):
-    user_to_delete = Users.query.get_or_404(id)
-    name = None
-    password_hash = None
-    form = Register()
-    try:
-        db.session.delete(user_to_delete)
-        db.session.commit()
-        flash("User Deleted Successfully!")
-        our_users = Users.query.order_by(Users.date_added)
-        return redirect(url_for('register'))
-    except:
-        flash("Whoops! There Was A Problem Deleting The User!")
-        return redirect(url_for('register'))
+    if id == current_user.id:
+        user_to_delete = Users.query.get_or_404(id)
+        name = None
+        password_hash = None
+        form = Register()
+        try:
+            db.session.delete(user_to_delete)
+            db.session.commit()
+            flash("User Deleted Successfully!")
+            our_users = Users.query.order_by(Users.date_added)
+            return redirect(url_for('register'))
+        except:
+            flash("Whoops! There Was A Problem Deleting The User!")
+            return redirect(url_for('dashboard'))
+    else:
+        flash("Sorry, You Can't Delete That User!")
+        return redirect(url_for('dashboard'))
     
 # Create Add Posts Page
 @app.route('/add-post', methods=['GET', 'POST'])
@@ -134,7 +154,7 @@ def add_post():
         db.session.add(post)
         db.session.commit()
         flash("Post Submitted Successfully!")
-        return redirect(url_for('add_post'))
+        return redirect(url_for('posts'))
     return render_template("add_post.html", form = form)
 
 # Create View Posts Page
@@ -157,7 +177,6 @@ def edit_post(id):
     form = PostForm()
     if form.validate_on_submit():
         post.title = form.title.data
-        # post.author = form.author.data
         post.slug = form.slug.data
         post.content = form.content.data
         # Update db
@@ -165,9 +184,8 @@ def edit_post(id):
         db.session.commit()
         flash("Post Updated Successfully!")
         return redirect(url_for('post', id=post.id))
-    if current_user.id == post.poster_id:
+    if current_user.id == post.poster_id or current_user == 1:
         form.title.data = post.title
-        # form.author.data = post.author
         form.slug.data = post.slug
         form.content.data = post.content
         return render_template('edit_post.html', form=form)
@@ -181,7 +199,7 @@ def edit_post(id):
 def delete_post(id):
     post_to_delete = Posts.query.get_or_404(id)
     id = current_user.id
-    if id == post_to_delete.poster.id:
+    if id == post_to_delete.poster.id or id == 1:
         try:
             db.session.delete(post_to_delete)
             db.session.commit()
@@ -196,6 +214,25 @@ def delete_post(id):
         flash("You're Not Authorized To Delete That Post!")
         posts = Posts.query.order_by(Posts.date_posted)
         return render_template("posts.html", posts=posts)
+
+# Pass Stuff To Navbar
+@app.context_processor
+def base():
+    form = SearchForm()
+    return dict(form=form)
+
+# Create Search Function
+@app.route('/search', methods=["POST"])
+def search():
+    form = SearchForm()
+    posts = Posts.query
+    if form.validate_on_submit():
+        # Get data from form submitted
+        post.searched = form.searched.data
+        # Query the database
+        posts = posts.filter(Posts.content.like('%' + post.searched + '%'))
+        posts = posts.order_by(Posts.title).all()
+        return render_template("search.html", form=form, searched=post.searched, posts=posts)
     
 @app.route('/admin')
 @login_required
